% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{sdf_intersects}
\alias{sdf_intersects}
\alias{sdf_contains}
\alias{sdf_within}
\alias{sdf_crosses}
\alias{sdf_covers}
\alias{sdf_covered_by}
\alias{sdf_equals}
\alias{sdf_disjoint}
\alias{sdf_touches}
\alias{sdf_overlaps}
\title{Spatial Predicate Functions
Predicates are optional, but enable the use of \code{sdf_join()}.}
\usage{
sdf_intersects(x, y, ...)

sdf_contains(x, y, ...)

sdf_within(x, y, ...)

sdf_crosses(x, y, ...)

sdf_covers(x, y, ...)

sdf_covered_by(x, y, ...)

sdf_equals(x, y, ...)

sdf_disjoint(x, y, ...)

sdf_touches(x, y, ...)

sdf_overlaps(x, y, ...)
}
\description{
Predicates should return a sparse matrix list representation. The elements the list
are an integer vector indicating the features where the predicate is \code{TRUE}.
The vector must be the \emph{row position}.
}
