% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{union_geometry}
\alias{union_geometry}
\alias{simplify_geometry}
\alias{buffer_geometry}
\alias{centroid}
\alias{convex_hull}
\alias{concave_hull}
\title{Primary Interface}
\usage{
union_geometry(x)

simplify_geometry(x, ...)

buffer_geometry(x, distance, ...)

centroid(x)

convex_hull(x)

concave_hull(x, concavity, ...)
}
\description{
These functions are optional but strongly suggested.
}
