% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf-join.R
\name{sdf_join}
\alias{sdf_join}
\title{Left join only}
\usage{
sdf_join(x, y, predicate, ..., suffix = c("_x", "_y"))
}
\arguments{
\item{predicate}{a function that returns a sparse matrix list representation}
}
\description{
Left join only
}
