% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-template.R
\name{use_sdf_template_class}
\alias{use_sdf_template_class}
\alias{use_sdf_template_zzz}
\title{Templates to extend sfd}
\usage{
use_sdf_template_class(
  class,
  save_as = paste0("R/sdf-compat-", class, ".R"),
  open = TRUE
)

use_sdf_template_zzz(class, save_as = "R/zzz.R", open = TRUE)
}
\arguments{
\item{class}{a scalar character of the class name of your geometry class}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
These templates can be used to extend sdf by implementing a new compatible
geometry vector column.
}
\details{
\itemize{
\item \code{use_sdf_template_class()} will create an R script with the scaffolding
needed to implement your geometry class to be compatible with \code{sdf}.
\item \code{use_sdf_template_zzz()} will create a \code{zzz.R} file that will export your
methods to be available to sdf.
}
}
